Add more granular engine config metrics
Creates a new metrics type which is closely aligned with osemgrep's
Engine_type.t which allows for more precise tracking of the engine
feature matrix. This is needed compared to a simple enumeration since we
have more orthogonal features than previously. For instance, we can now
perform analysis on at least two axes (within pro):
    (1) secrets valdiation [enabled <-> disabled];
    (2) dataflow [none <-> ... <-> interfile]
Creating a new enum entry for each possible permutation is not
futureproof (and is ugly), so we need a more complex type here.