Make git blob optional
Really this should probably not be but needed for backwards
compatibility, even though we haven't sent historical info yet.

Might be reasonable to allow not sending it though, even for updated
things, if the object we want to reporting something out is really a
submodule or something? I don't think a non-blob object makes sense
today though.