Revert "quick-fix: Assign .h extension only to C++ (not to C) (#233)"
This reverts commit b830f0147555b1fab39a311a58f4a5f93c077dd8.

I was working under the wrong assumptions. I believe that what it
actually happens when Semgrep encounters a .h file is that it
considers it both a C and a C++ file. Thus the .h file will be scanned
twice (as C, and as C++ code, independently). If the .h file is a C++
file, when scanned as C, it may produce a bunch of parse errors (e.g.
because there is a class declaration in it) which then results in those
"N skipped lines" being reported by --verbose. But when scanned as C++
the same file seems to work just fine.

The C++ scan should not be negatively impacted, but the user is being
mislead.

I still think that the best solution could be to make C an alias for C++
and handle all C and C++ files together. Otherwise the CLI would need to
be smarter when doing file targeting and have some heuristic to avoid
classifying C++ headers as C code.