Extend the definition of 'analyzer' in Semgrep rules (#286)
This PR extends the Semgrep rule schema in accordance with the work I
did on the OCaml side. The PR is
https://github.com/semgrep/semgrep-proprietary/pull/1641 and I'm going
to push my work to it very soon.

This PR also adds JSON Schema validation tests. They're useful to check
that the JSON Schema is valid, that it validates valid Semgrep rules,
and rejects invalid Semgrep rules - without having to rely on another
git repo to test this.

test plan: `make test`

This will run `check-jsonschema`. `make setup` will ask the user to
install it if it's missing.

- [x] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [x] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
generated
	  by Semgrep 1.17.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades