chore(dep-resolution): update dep-resolution RPC interface to support lockfiles (#312)
Updates the RPC interface used to resolve dependencies in OCaml to
support
lockfiles as well as manifests. This will allow us to begin porting
lockfile
parsers to Ocaml as we move forward.

These changes also make it possible to report errors that do not cause
dependency resolution to fail entirely.

These changes are _not_ backwards compatible. However, they only affect
types
that are used within the CLI, from the Python code to the OCaml code, so
as long
as we are careful with releasing the changes to the CLI, it should be
safe.

- [x] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [ ] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
generated
	  by Semgrep 1.17.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades

This must be merged in sync with
https://github.com/semgrep/semgrep-proprietary/pull/2571 since it has
breaking changes.