Remove semgrep_version from project_metadata and meta from scan_request
The backend should not use those fields anymore so
we can remove them. Note that even if old versions
of semgrep are still producing those fields, it is fine
because when the backend is consuming the JSON, any unknown
field is skipped and those fields should not be used anymore.

test plan:
see related PR in semgrep