chore(sca): add flag to disable subproject change detection (#335)
Adds a new flag to pass from the app to the CLI that forces the CLI to
resolve
dependencies for all subprojects, all the time (even for diff scans).
Some
customers want this so that we get a full picture of their dependencies
for
every branch.

- [x] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [x] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
	  generated by Semgrep 1.50.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades
	  Note that the types related to the semgrep-core JSON output or the
	  semgrep-core RPC do not need to be backward compatible!