feat(scrat): symbol analysis (#341)
This PR adds "symbol analysis", a dump of the third-party symbols used
by a project, to the scan results for Semgrep.

This will enable the `scrat` project.

- [X] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [X] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
	  generated by Semgrep 1.50.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades
	  Note that the types related to the semgrep-core JSON output or the
	  semgrep-core RPC do not need to be backward compatible!