chore(sca): add rpc to match subprojects (#364)
Adds an RPC type that allows Python to ask Ocaml to match subprojects
based on a list of dependency source files.
This will allow us to switch to ocaml-based subproject matching
independently from the rest of file targeting.

- [x] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [x] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
	  generated by Semgrep 1.50.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades
	  Note that the types related to the semgrep-core JSON output or the
	  semgrep-core RPC do not need to be backward compatible!