Add instructions to derive comparison functions for sorting purposes (#382)
This adds a lot of ppx_deriving.ord annotations allowing us to sort
results deterministically as opposed to the lazy approach where we'd
omit fields when sorting a list of results.

No types have changed.

- [x] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [x] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
	  generated by Semgrep 1.50.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades
	  Note that the types related to the semgrep-core JSON output or the
	  semgrep-core RPC do not need to be backward compatible!