feat: add fips mode flag to scan_configuration, core_match (#386)
As discussed in SAF-2057, semgrep_app will use this flag to indicate to
the CLIs when FIPS-compliant hashing is required for, in particular,
match-based IDs.

test plan: this has been run locally with semgrep-app and semgrep-core
(details
[here](https://github.com/semgrep/semgrep-proprietary/pull/4080).)

Since the change to the IDL files is only an optional field, we should
be backward-compatible with previous versions.

The accompanying changes for
[semgrep-proprietary](https://github.com/semgrep/semgrep-proprietary/actions/runs/15693843980/job/44214806689?pr=4080)
and [semgrep-app](https://github.com/semgrep/semgrep-app/pull/21648)
will not be merged until this PR is merged.


- [x] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [x] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
	  generated by Semgrep 1.50.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades
	  Note that the types related to the semgrep-core JSON output or the
	  semgrep-core RPC do not need to be backward compatible!
- [x] Any accompanying changes in `semgrep-proprietary` are approved and
ready to merge once this PR is merged