remove duplicate FIPS flag (#388)
The original implementation of SAF-2057 tracked `fips_mode` in
`core_match_extra`, but this was removed during the review process; the
interfaces submodule was only updated locally but the change was never
pushed up. This patch does so.

This flag was never used by any semgrep releases so is safe to remove.

- [x] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [x] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
	  generated by Semgrep 1.50.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades
	  Note that the types related to the semgrep-core JSON output or the
	  semgrep-core RPC do not need to be backward compatible!
- [x] Any accompanying changes in `semgrep-proprietary` are approved and
ready to merge once this PR is merged