Keep track of ppaths after target discovery so as to allow rule filtering (#375)
The ppath of a target file is its path relative to the project root
identified on the OCaml side during target discovery (`Find_targets`
module). It is now returned to pysemgrep so that rule filtering based on
the paths.include/exclude patterns can take place correctly (as per the
Semgrepignore/Gitignore spec) on the ppaths in Python and then again in
OCaml when the targets are sent to semgrep-core.

Goes with https://github.com/semgrep/semgrep-proprietary/pull/3859

- [x] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [x] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
	  generated by Semgrep 1.50.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades
	  Note that the types related to the semgrep-core JSON output or the
	  semgrep-core RPC do not need to be backward compatible!