perf(cli): support compressing metadata (#397)
### Description

To reduce CLI scan startup times, we are going to compress the scan
config before sending it from the backend to the CLI. In order to be
backwards compatible, we are adding an optional `compress_config`
parameter so the CLI clients can opt in to response compression when
they upgrade.

### Testing

Working on a local semgrep CLI -> local Semgrep app test for this
change.


- [x] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [x] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
	  generated by Semgrep 1.50.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades
	  Note that the types related to the semgrep-core JSON output or the
	  semgrep-core RPC do not need to be backward compatible!
- [ ] Any accompanying changes in `semgrep-proprietary` are approved and
ready to merge once this PR is merged