chore(cli): revert unneeded new parameter (#400)
### Description

Reverting https://github.com/semgrep/semgrep-interfaces/pull/397 because
we have a better solution that does not need this parameter.

I talked with @salolivares and we think this might be fine even though
it's not backwards compatible because neither the server nor the CLI
ever actually started using this parameter. But I'm actually not 100% on
if it's ok to make a non backwards compatible change even if the
parameter was never set or read by anyone so I'd be interested to hear
from someone who could confirm that.


- [x] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [x] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
	  generated by Semgrep 1.50.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades
	  Note that the types related to the semgrep-core JSON output or the
	  semgrep-core RPC do not need to be backward compatible!
- [ ] Any accompanying changes in `semgrep-proprietary` are approved and
ready to merge once this PR is merged
- https://github.com/semgrep/semgrep-app/pull/23008
- https://github.com/semgrep/semgrep-proprietary/pull/4481