chore(sca): add local-builds flag to dependency resolution rpc interface (#402)
We were previously relying on the caller of this RPC to only call it
when
`allow-local-builds` was enabled. But as we start to do more non-dynamic
lockfile parsing in OCaml, this becomes less feasible and more prone to
mistakes. This PR passes the flag through the RPC call so that the OCaml
code
can decide itself which resolution method to use.

- [x] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [x] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
	  generated by Semgrep 1.50.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades
	  Note that the types related to the semgrep-core JSON output or the
	  semgrep-core RPC do not need to be backward compatible!
- [x] Any accompanying changes in `semgrep-proprietary` are approved and
ready to merge once this PR is merged