chore(lfs): add resolution error for missing resource (#420)
A common error class for us in lockfileless is inaccessible resources
that are
needed for the build. This is typically caused by incomplete private
package
manager configuration, but could theoretically also be caused by a
package
manager misconfiguration. I believe that in all cases, there is not much
we can
do on our end. So this PR creates a new type of resolution error to
capture
these.

- [x] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [x] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
	  generated by Semgrep 1.50.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades
	  Note that the types related to the semgrep-core JSON output or the
	  semgrep-core RPC do not need to be backward compatible!
- [x] Any accompanying changes in `semgrep-proprietary` are approved and
ready to merge once this PR is merged