Format comments to make them visible in the generated OCaml code (#422)
I reformatted the main ATD file so as to propagate the documentation
into the generated OCaml code. Note that we could have the same for
Python and TypeScript but we'd have to implement it in atdpy and atdts.

I also added the generated files `semgrep_output_t.ml[i]` to this repo.
They're needed if we don't want to regenerate them like we've been doing
so far in semgrep. I'll make a semgrep PR that symlinks to the OCaml
files so that we can jump to to these type definitions like for any
other OCaml code.

This PR does not change type definitions.

- [x] I ran `make setup && make` to update the generated code after
editing a `.atd` file (TODO: have a CI check)
- [x] I made sure we're still backward compatible with old versions of
the CLI.
For example, the Semgrep backend need to still be able to *consume* data
	  generated by Semgrep 1.50.0.
See
https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades
	  Note that the types related to the semgrep-core JSON output or the
	  semgrep-core RPC do not need to be backward compatible!
- [x] Any accompanying changes in `semgrep-proprietary` are approved and
ready to merge once this PR is merged