chore(mcp): add TP/FP/skip tracking fields (#423)
This PR works together with [this PR](https://github.com/semgrep/semgrep-proprietary/pull/5091) in `semgrep-proprietary`​. The PR in the main repo updates the MCP server to let users flag TP/FP on findings found in an MCP scan. It also send metrics based on user feedback on whether the finding is a TP or FP, which is why we are updating the interface here.

- [x] I ran `make setup && make` to update the generated code after editing a `.atd` file (TODO: have a CI check)
- [x] I made sure we're still backward compatible with old versions of the CLI.
For example, the Semgrep backend need to still be able to _consume_ data
generated by Semgrep 1.50.0.
See https://atd.readthedocs.io/en/latest/atdgen-tutorial.html#smooth-protocol-upgrades
Note that the types related to the semgrep-core JSON output or the
semgrep-core RPC do not need to be backward compatible!
- [x] Any accompanying changes in `semgrep-proprietary` are approved and ready to merge once this PR is merged